requirements.txt
dlib==19.24.2
numpy==1.23.5
opencv-python-headless==4.8.0.74
Pillow==9.5.0
scipy==1.11.1
torch==2.0.1
torchvision==0.15.2
